+++ tests/tt_metal/tt_metal/perf_microbenchmark/routing/test_tt_fabric_sanity.cpp
@@ -53,6 +53,7 @@
 #include "umd/device/types/xy_pair.h"
 #include <tt-metalium/utils.hpp>
 #include "tt_metal/fabric/fabric_host_utils.hpp"
+#include "tt_metal/fabric/hw/inc/fabric_routing_mode.h"
 
 using std::vector;
 using namespace tt;
