+++ tt_metal/fabric/fabric_mux_config.hpp
@@ -9,7 +9,7 @@
 #include <tt-metalium/fabric.hpp>
 #include <tt-metalium/erisc_datamover_builder.hpp>
 #include "impl/context/metal_context.hpp"
-#include "tt_metal/fabric/fabric_host_utils.hpp"
+#include "tt_metal/fabric/fabric_context.hpp"
 
 namespace tt::tt_fabric {
